Upstream recent downstream changes in envsetup.sh.
A superset of this change was needed to fix a bug downstream. Note that the bug
doesn't exist upstream since it was caused by a piece of code not upstreamed.

Therefore the only purpose of this change is to decrease the upstream vs
downstream diff.


Review URL: https://chromiumcodereview.appspot.com/11365147

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ed5f0199f1c787712f326b3680373aa8df97277a