Modify the Android test running scripts so they can be used from WebKit
In order to avoid duplicating most of the run-tests-on-devices logic, make
sure that the Android scripts are able to be usable from a WebKit checkout.
Two flags have been added for this: --webkit, which is necessary as a file's
path on the host and device are not per se equal anymore, and --out-directory,
accepting a path as it's not always in the Chromium source directory.

BUG=


Review URL: https://chromiumcodereview.appspot.com/11312239

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d933f5e9992b96002e76dd11b48d2e3797c016db