android: Print error when trying to build on 32-bit host.
This patch changes build/android/envsetup.sh to detect that the user
is running on a 32-bit host system, and print an error message
in this case, since a full build will fail (linker runs out of memory,
because the generated files are too large).

This can be overriden with the new --try-32 flag, in case someone
really wants to try a build. This would only be useful if one
needs to rebuild a "small" target, not do a complete build.

BUG=


Review URL: https://chromiumcodereview.appspot.com/11366243

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8f773620cac4539d56c1dc4f0c09076ffefb5bc9