Add an accessibility audit for WebUI pages.
This change adds an accessibility audit which runs just before tearDown() on any WebUI browser test that uses the chrome/test/data/webui/test_api.js framework.

As of this change, the test will run by default for each test case, but not cause any errors in the case of accessibility issues. As we fix accessibility issues, we can set the a11yIssuesAreErrors flag on a per-test and per-fixture basis, to catch regressions. Once the majority of WebUI pages are passing the accessibility audit, and we have a mechanism in place for ignoring false positives, we will transition to treating accessibility issues as errors by default.

http://www.chromium.org/developers/accessibility/webui-accessibility-audit has more information.

BUG=162740

Review URL: https://chromiumcodereview.appspot.com/11363170

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a8f7362f3dd77a2c8645ed1c3ee8fd9ee10797a0