Rearrange gyp files for Message Center
In order to facilitate the move to MessageCenterNotificationManager, we
need to be able to determine when message_center.gyp is being included
in the build.  This change defines ENABLE_MESSAGE_CENTER which can be
used for conditional compilation.

r=stevenjb@chromium.org
TBR=ben@chromium.org

BUG=168605

Review URL: https://chromiumcodereview.appspot.com/11791004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9b5dbf76bd8f12781fa9fb1202509de9c315aa35