Android: Propagate timeouts when sharding(and retrying) tests.
We use TestResult.GetAllBroken to identify and retry tests.
However, this list is empty if a test times out (and the device is
responsive). In this case we should propagate the timed_out info to the
final result, which will report a STEP_FAILURE based on this
information.

Ex. bad run:
http://build.chromium.org/p/chromium.linux/builders/Android%20Tests%20%28dbg%29/builds/6093/steps/sandbox_linux_unittests/logs/stdio

BUG=169064
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/11833017

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c92d09e892daa221fe2ad42f64c314c1a8dffcac