Fix PRESUBMIT.py for build/android/buildbot
The code was making a call to android_commands.py, which
relied on a copy of adb existing in the path.  The CQ
bot does not have third_party/android_tools, and in general
it seems advantageous to not require a full checkout for
syntax checks.

NOTRY=True

Review URL: https://chromiumcodereview.appspot.com/11883019

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c94ae6bb71cb75306117eab5dfcca3066f5c3625