Revert "Set msvs_cygwin_shell default to 0"
This reverts commit 27a1f2c62be751eb8926f4792cdd3b5d4baac310.

I suspect this is causing issues for the Win7 Tests (1) bot.
Something went wrong that causes a lot of browser_tests to fail.
I suspect that a step is running incorrectly on the build side
which causes the packaged build to not work on the testers.

http://build.chromium.org/p/chromium.win/builders/Win7%20Tests%20%281%29/builds/19113

TBR=iannucci@chromium.org

Review URL: https://codereview.chromium.org/12041037

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5f75fb60507487b336985340ad54e1df52ee614e