Clean up control of perf governor
The perf governor was previously controlled by the PerfTestSetup class,
which set the perf governor to 'performance' mode in Startup() and
reset it to its original value on TearDown(). This had a number of
problems:
- We ended up with TearDown() being used in place of SetUp() when
starting tests where we didn't want performance mode. This looks
strange.
- There was no explicit way of setting the device's default
performance mode. We would call TearDown() and hope for the best. This
would not work if the device was already in performance mode at the
start of the test. We think this is sometimes happening on the
performance bots.
- SetUp() also cleared the caches, which should have no direct
connection with setting the performance mode.

This patch creates two new classes, PerfControl and CacheControl, to
manage the perf governor and cache. These replace PerfTestSetup.

PerfTestSetup has not yet been deleted to avoid the need for simultaneous
commits on the upstream and downstream repositories.


Review URL: https://chromiumcodereview.appspot.com/12720010

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5c77935598556ee89d445467802adbb6237d53b5