Failing layout tests on Android shouldn't fail the whole bot.
Output a warning instead of a failure when the layout test command
fails.  We'll revert this when we've got a stable base, and it will
be easier for others to track webkit_unit_tests and TestWebKitAPI
failurse until that happens.

BUG=218257

Review URL: https://chromiumcodereview.appspot.com/12709003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8feb367a9804f6eccd09531f28a04f1d7caa6c3c