Fix proguard
Proguard should only be run in Release mode (in gyp, proguard_enabled is
not set based on the configuration). This is the way that we had
been determining whether or not to run proguard before recent Ant
refactoring.

Also, proguard needs the path to the android.jar


BUG=224079


Review URL: https://chromiumcodereview.appspot.com/12922004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3adb9b2d5719b3afbbd2b7b92fa534c7ceb304e1