Reorder platforms in build_config to fix Android on Mac.
The Android port relies on the global define ANDROID to identify that
we are crosscompiling for Android, but __APPLE__ was being checked first
in build_config.h, and so builds on a Mac sometimes ended up with the
wrong OS_* macro defined. Reorder the list so that ANDROID is tested
first, as it will only be set manually when crosscompiling (__APPLE__ is
set internally by the compiler).

BUG=


Review URL: https://chromiumcodereview.appspot.com/13541003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c57a3708429fb283feccd580e691056b9ca4c4d4