[Android] Extend setup timeout for content_browsertests.
content_browsertests were timing out on the Android FYI bots since it
took more than 7 minutes to push the test files to the device. This
extends the timeout to 20 minutes, as a stopgap until we can reduce
the number of files that need to be pushed.

BUG=138275


Review URL: https://chromiumcodereview.appspot.com/13674021

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 55aa82aa903ea06283bf53d3038f112f8c8b42c9