Make IsHostPortUsed() handle ports bound by foreign UIDs.
lsof -np is not able (permission denied when running as non-root to be exact)
to retrieve file descriptor information for processes running under a different
UID. This can make IsHostPortUsed() return false when the provided port is
actually used.
Trying to connect to a port rather than binding it should not introduce any
side-effect on the system that could affect next runs as the TODO suggested. It
is also always sane not to depend on an external process (lsof) output format.

BUG=229685
R=bulach@chromium.org

Review URL: https://codereview.chromium.org/14567016

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f1ed7b6d39a6ad8bb6c71e4c4d28e391deb72bc3