Remove ENABLE_MESSAGE_CENTER
Next step on re-factoring Notifications. This will bring the MessageCenter classes, most important message_center::Notification, into build on all systems, including those where the MessageCenter is not yet appearign in UI. This will allow to start using this Notification class in client code and remove the old Notification class defined in chrome/browser/notificaitons/notification.h

That will allow the clients that already use Rich Notifications to use richer data type support, for example supply an image bits for a Notification (as in Snapshot notifications).

This also removes a lot of compile-time @ifdefs and replaces them with checking a runtime flag which we already have anyways.

On Android and iOS, the MessageCenter is not compiled in, for the size concerns and uncertain story for notifications in general - the existing ENABLE_NOTIFICATIONS define is used for that.

BUG=174164

Review URL: https://chromiumcodereview.appspot.com/15025002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bce52eaa760e4a43b96c69e92b18af204c04546e