Fix test for libudev1
The test for libudev1/libudev0 fails when an sdk has installed
a dependency on libudev1, which will go away on recompile, on
a system with only libudev0.  The dependency created an empty
virtual package that passes the "show" test, but cannot be installed.

BUG=243398
TEST=local

Review URL: https://chromiumcodereview.appspot.com/15911004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 51c29c2518fb62377aa82c67d964fe63edacfaf8