Switch all windows bots to use Aura.
So we can have the perf bots run with aura for a while. Depending how red
the waterfall we could leave it on for a long time.

This has happened before and it will happen again.
See for example r204698

BUG=259185
TEST=none
TBR=scottmg

Review URL: https://codereview.chromium.org/18434006

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6e5936cd6be71bb95fde1f8f4da039ea16d66948