cc: Consider scroll offset in CalcDrawProperties
Previously, any updates to scroll offset on the main thread updated both
scroll offset and layer position.  CalcDrawProperties only used position
and ignored scroll offset.  This patch changes that to consider scroll
offset equally with scroll delta when computing a layer's tranform.

Although this patch doesn't do anything about it yet, the end result of
this disentangling is that it will become possible to early out in
Layer::SetScrollOffset and not require a SetNeedsCommit when Blink
updates the scroll offset to the value that it already is on the
compositor thread.

This is step 3 of a patch series to disentangle layer position and
scroll offset.  This patch depends on both
https://codereview.chromium.org/18405003/ and
https://codereview.chromium.org/18187004/.

R=danakj@chromium.org
BUG=256381

Review URL: https://chromiumcodereview.appspot.com/18400003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4ec78f81c97f9aaed2fd052182c208ad90b4988a