Add test_server_setup.py and add support for additional flags.
Whenever a host driven test need to setup a server as part of a test,
the command line needs of the application to be instrumented needs to
change.

This CL adds a helper class for setting up a test server, and also adds
support for tests to setup additional flags (typically the URL for how
to connect to the test server).

Depends on: https://codereview.chromium.org/23467004/

BUG=272584
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/23726004

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5d90fbebb5095e403cf753af35914efebf43efe4